Preview scales badly (fixes #22).

+++ intervalmusiccompositor.fxUi/src/main/java/ch/retorte/intervalmusiccompositor/ui/graphics/BarChart.java
@@ -150,7 +150,7 @@ public class BarChart {
 
 
         if (has(breakPattern)) {
-          p = p + (int) (breakPattern.get(j % breakPattern.size()) * scale);
+          p = p + (breakPattern.get(j % breakPattern.size()) * scale);
         }
       }
     }
